pyproject.toml
[project]
name = "nagchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=2.19.0",
    "faiss-cpu>=1.7.4",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "numpy>=1.26.4",
    "ollama>=0.1.9",
    "pydantic>=2.7.0",
    "pypdf2>=3.0.1",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.1",
    "ragas>=0.1.9",
    "sentence-transformers>=2.7.0",
    "streamlit>=1.33.0",
    "tiktoken>=0.7.0",
    "torch>=2.2.0",
    "transformers>=4.41.0",
]
